refactor(mesh): rename environment-variables.ts to config.ts
